pyproject.toml
[tool.poetry]
name = "zidship"
version = "0.1.0"
description = ""
authors = ["MohamedSamir <mhmdsamir92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
reportlab = "^3.6.12"
drf-spectacular = "^0.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
